refactor(admin): extract System Instructions, Sources, and Settings sections from AppFormEditor (#1781)
Follow-up to #1781, continuing the series of small extraction PRs already
opened against it (#1942, #2044, #2046, #2048, #2049, #2050, #2051, #2054,
#2060). Extracts three of the remaining unclaimed sections into standalone
components under client/src/features/admin/components/app-form/:

- SystemInstructionsSection.jsx — system prompt, message placeholder, and
  prompt template editors for chat apps
- SourcesConfigSection.jsx — data source picker + sources summary card
- SettingsConfigSection.jsx — per-chat settings toggles (model, temperature,
  output format, chat history, style)

Each component receives (app, onChange, t) and owns its own local update
handler, matching the convention established by the prior PRs in this
series. Pure extraction — no behavior change intended.

AppFormEditor.jsx shrinks from 2,902 to 2,667 lines.
